requirements.txt
# Core
schedule==1.2.1
python-dotenv==1.0.0
supabase==2.9.1
anthropic==0.43.0
requests==2.31.0

# Collectors
feedparser==6.0.10
pytrends==4.9.2
praw==7.7.1
google-api-python-client==2.118.0
telethon==1.34.0

# Processors
pysentimiento==0.7.2
bertopic==0.16.2
sentence-transformers==2.7.0

# Dashboard
streamlit==1.31.1
plotly==5.18.0
pandas==2.1.4
